+++ dp_wizard/utils/code_generators/analyses/histogram/no-tests/_histogram_config.py
